Created MongoSequence class for autogenerate id
Created a new MongoSequenceConfig class, along with a player listener
and dice game listener to auto incremente the object ids whenever a new
player is created, as well as a new dice game that is added to the
player database. mongo doesnt recognize simple integers and has its own
id system within. hence a sequence class and database is needed to be
created as well as a serivce to implement the logic to it.